fix(types+model): raw-order sealed + full-width retry bounds -- review round 13 (GH #476 Change 3)
Two P1s from the round-13 review of PR #481:

1. The legacy encoder sorts sealed loci by RAW canonical name and
   demangles only while serializing -- display values in raw
   order. The projection's BTreeSet-of-displays re-sorted by alias
   spelling. entities.loci is already canonically raw-sorted, so
   the projection now collects sealed displays into a Vec without
   re-sorting. Pinned with two imports whose raw and alias orders
   deliberately disagree (raw __lib_a_pack_Z < __lib_b_pack_A,
   displays z::Z then a::A -- non-lexical display order is the
   correct output).

2. SupervisionPolicy.retry_bound was Option<u32> while the legacy
   artifact keeps the literal as written (i64) -- 'for 4294967296'
   is check-clean and the 'as u32' cast projected it as 0. The
   field is i64 through schema, builder, and projection; no source
   rejection (that would be a language change and contradict this
   PR's no-user-visible-change goal).

Both fixes negative-tested. 1235/1235.

+++ crates/hale-model/src/relation.rs
@@ -275,7 +275,11 @@ pub struct TopicBinding {
 #[derive(Clone, PartialEq, Eq, Debug)]
 pub struct SupervisionPolicy {
     pub ops: Vec<String>,
-    pub retry_bound: Option<u32>,
+    /// The literal as WRITTEN — i64 like the legacy artifact, which
+    /// serializes any int literal the parser accepted (`for
+    /// 4294967296` is check-clean; a narrower type here silently
+    /// truncated it — review round 13).
+    pub retry_bound: Option<i64>,
 }
 
 /// What an `on_failure` handler supervises. Usually a declared
